pyproject.toml
[project]
name = "algotrader"
version = "0.1.0"
description = "Adaptive multi-strategy algorithmic trading system for Alpaca Markets"
requires-python = ">=3.11"
dependencies = [
    "alpaca-py>=0.28.0",
    "ib_async>=2.0.1",
    "nest_asyncio>=1.6.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "scipy>=1.12.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "apscheduler>=3.10.0",
    "structlog>=24.1.0",
    "streamlit>=1.31.0",
    "pyyaml>=6.0.0",
    "pytz>=2024.1",
    "python-dotenv>=1.0.0",
    "statsmodels>=0.14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
